#2 Criação e envio da mensagem por email
