+++ app/DI/ServicesConfig.cs
@@ -7,9 +7,7 @@ namespace app.DI
     {
         public static void AddConfigServices(this IServiceCollection services, IConfiguration configuration)
         {
-            services.AddScoped<IUsuarioService, UsuarioService>();
             services.AddScoped<ISolicitacaoAcaoService, SolicitacaoAcaoService>();
-            services.AddScoped<IEmailService, EmailService>();
         }
     }
 }
