--- service/Interfaces/IEmailService.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace service.Interfaces
-{
-    public interface IEmailService
-    {
-        public void EnviarEmail(string emailDestinatario, string assunto, string corpo);
-    }
-}
