+++ service/SolicitacaoAcaoService.cs
@@ -14,16 +14,48 @@ namespace service
 {
     public class SolicitacaoAcaoService : ISolicitacaoAcaoService
     {
-        private readonly IEmailService emailService;
-        public SolicitacaoAcaoService(IEmailService emailService)
+        public SolicitacaoAcaoService()
         {
-            this.emailService = emailService;
+
         }
         public void EnviarSolicitacaoAcao(SolicitacaoAcaoDTO solicitacaoAcaoDTO)
         {
-            emailService.EnviarEmail("dnit@gmail.com", "Solicitação de Serviço", "");
+            string mensagem = $"Nova solicitação de ação em escola\n\n" +
+                            $"Escola: {solicitacaoAcaoDTO.Escola}\n" +
+                            $"Nome do Solicitante: {solicitacaoAcaoDTO.NomeSolicitante}\n" +
+                            $"Vínculo com a escola: {solicitacaoAcaoDTO.VinculoEscola}\n" +
+                            $"Email: {solicitacaoAcaoDTO.Email}\n" +
+                            $"Telefone: {solicitacaoAcaoDTO.Telefone}\n" +
+                            $"Ciclos de ensino: {solicitacaoAcaoDTO.CiclosEnsino}\n" +
+                            $"Quantidade de alunos{solicitacaoAcaoDTO.QuantidadeAlunos}\n" +
+                            $"Observações: {solicitacaoAcaoDTO.Observacoes}\n";
+            string emailDestinatario = Environment.GetEnvironmentVariable("EMAIL_DNIT");
+            EnviarEmail(emailDestinatario, "Solicitação de Serviço", mensagem);
 
 
+
+        }
+        public void EnviarEmail(string emailDestinatario, string assunto, string corpo)
+        {
+
+            MailMessage mensagem = new MailMessage();
+
+            string emailRemetente = Environment.GetEnvironmentVariable("EMAIL_SERVICE_ADDRESS");
+            string senhaRemetente = Environment.GetEnvironmentVariable("EMAIL_SERVICE_PASSWORD");
+
+            mensagem.From = new MailAddress(emailRemetente);
+            mensagem.Subject = assunto;
+            mensagem.To.Add(new MailAddress(emailDestinatario));
+            mensagem.Body = corpo;
+
+            var clienteSmtp = new SmtpClient("smtp-mail.outlook.com")
+            {
+                Port = 587,
+                Credentials = new NetworkCredential(emailRemetente, senhaRemetente),
+                EnableSsl = true,
+
+            };
+            clienteSmtp.Send(mensagem);
         }
     }
 }
